--- tests/Acceptance.suite.yml
@@ -1,16 +1,0 @@
-# Codeception Test Suite Configuration
-#
-# Suite for acceptance tests.
-# Perform tests in browser using the WebDriver or PhpBrowser.
-# If you need both WebDriver and PHPBrowser tests - create a separate suite.
-
-actor: AcceptanceTester
-modules:
-  enabled:
-    - PhpBrowser:
-        url: http://localhost/myapp
-# add Codeception\Step\Retry trait to AcceptanceTester to enable retries
-step_decorators:
-  - Codeception\Step\ConditionalAssertion
-  - Codeception\Step\TryTo
-  - Codeception\Step\Retry
