--- tests/Functional.suite.yml
@@ -1,12 +1,0 @@
-# Codeception Test Suite Configuration
-#
-# Suite for functional tests
-# Emulate web requests and make application process them
-# Include one of framework modules (Symfony, Yii2, Laravel, Phalcon5) to use it
-# Remove this suite if you don't use frameworks
-
-actor: FunctionalTester
-modules:
-  enabled:
-  # add a framework module here
-step_decorators: ~
