+++ tests/Support/IntegrationTester.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Tests\Support;
 
+use Codeception\Actor;
+
 /**
  * Inherited Methods
  * @method void wantTo($text)
@@ -19,9 +21,9 @@ namespace Tests\Support;
  *
  * @SuppressWarnings(PHPMD)
 */
-class AcceptanceTester extends \Codeception\Actor
+class IntegrationTester extends Actor
 {
-    use _generated\AcceptanceTesterActions;
+    use _generated\IntegrationTesterActions;
 
     /**
      * Define custom actions here
